fix: simplify Stellar trait and remove auth methods
